Move all RAS-RM integration code into its own package
