+++ src/main/java/uk/gov/ons/ssdc/caseprocessor/messaging/NewCaseReceiver.java
@@ -17,8 +17,8 @@ import uk.gov.ons.ssdc.caseprocessor.model.dto.EventDTO;
 import uk.gov.ons.ssdc.caseprocessor.model.dto.NewCase;
 import uk.gov.ons.ssdc.caseprocessor.model.repository.CaseRepository;
 import uk.gov.ons.ssdc.caseprocessor.model.repository.CollectionExerciseRepository;
+import uk.gov.ons.ssdc.caseprocessor.rasrm.service.RasRmCaseNotificationEnrichmentService;
 import uk.gov.ons.ssdc.caseprocessor.service.CaseService;
-import uk.gov.ons.ssdc.caseprocessor.service.RasRmCaseNotificationEnrichmentService;
 import uk.gov.ons.ssdc.caseprocessor.utils.CaseRefGenerator;
 import uk.gov.ons.ssdc.common.model.entity.Case;
 import uk.gov.ons.ssdc.common.model.entity.CollectionExercise;
