+++ src/main/java/uk/gov/ons/ssdc/caseprocessor/rasrm/model/dto/RasRmCaseGroupDTO.java
@@ -1,9 +1,9 @@
-package uk.gov.ons.ssdc.caseprocessor.model.dto;
+package uk.gov.ons.ssdc.caseprocessor.rasrm.model.dto;
 
 import java.util.UUID;
 import lombok.Data;
 
 @Data
-public class RasRmPartyLinkDTO {
+public class RasRmCaseGroupDTO {
   private UUID collectionExerciseId;
 }
