+++ src/main/java/uk/gov/ons/ssdc/caseprocessor/schedule/CaseClassifier.java
@@ -1,12 +1,11 @@
 package uk.gov.ons.ssdc.caseprocessor.schedule;
 
-import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-import uk.gov.ons.ssdc.caseprocessor.client.RasRmPartyServiceClient;
+import uk.gov.ons.ssdc.caseprocessor.rasrm.service.RasRmSampleSummaryCollexLinkService;
 import uk.gov.ons.ssdc.common.model.entity.ActionRule;
 
 @Component
@@ -15,12 +14,13 @@ public class CaseClassifier {
       Set.of("rasRmSampleSummaryId", "rasRmCollectionExerciseId");
 
   private final JdbcTemplate jdbcTemplate;
-  private final RasRmPartyServiceClient rasRmPartyServiceClient;
+  private final RasRmSampleSummaryCollexLinkService rasRmSampleSummaryCollexLinkService;
 
   public CaseClassifier(
-      JdbcTemplate jdbcTemplate, RasRmPartyServiceClient rasRmPartyServiceClient) {
+      JdbcTemplate jdbcTemplate,
+      RasRmSampleSummaryCollexLinkService rasRmSampleSummaryCollexLinkService) {
     this.jdbcTemplate = jdbcTemplate;
-    this.rasRmPartyServiceClient = rasRmPartyServiceClient;
+    this.rasRmSampleSummaryCollexLinkService = rasRmSampleSummaryCollexLinkService;
   }
 
   public void enqueueCasesForActionRule(ActionRule actionRule) {
@@ -29,34 +29,11 @@ public class CaseClassifier {
         .getSurvey()
         .getSampleDefinitionUrl()
         .endsWith("business.json")) {
-      Object metadataObject = actionRule.getCollectionExercise().getMetadata();
-
-      if (metadataObject == null) {
-        throw new RuntimeException(
-            "Unexpected null metadata. Metadata is required for RAS-RM business.");
-      }
-
-      if (!(metadataObject instanceof Map)) {
-        throw new RuntimeException(
-            "Unexpected metadata type. Wanted Map but got "
-                + metadataObject.getClass().getSimpleName());
-      }
-
-      Map metadata = (Map) metadataObject;
-
-      if (!metadata.keySet().containsAll(MANDATORY_COLLEX_METADATA)) {
-        throw new RuntimeException("Metadata does not contain mandatory values");
-      }
-
-      UUID rasRmSampleSummaryId = UUID.fromString((String) metadata.get("rasRmSampleSummaryId"));
-      UUID rasRmCollectionExerciseId =
-          UUID.fromString((String) metadata.get("rasRmCollectionExerciseId"));
-
       // This only needs to be done once, for efficiency, but it's pretty horrible having to hack
       // it in right here. In the ideal world, there would be a more elegant place to put this
       // in the code... but its horribleness is largely due to how difficult RAS-RM APIs are
-      rasRmPartyServiceClient.linkSampleSummaryToCollex(
-          rasRmSampleSummaryId, rasRmCollectionExerciseId);
+      rasRmSampleSummaryCollexLinkService.linkSampleSummaryToCollex(
+          actionRule.getCollectionExercise());
     }
 
     UUID batchId = UUID.randomUUID();
